applications/yieldmonproc/requirements.txt
cherrypy
mako
requests
simplejson
sh
python-dateutil